requirements.txt
Flask
flask_restful
mysql
pycryptodome
requests
datetime
jwcrypto
mysql-connector-python